removing old legacy block as thought
